Fixing periodic boundary conditions

+++ src/Grid/grid_dist_id_unit_test.hpp
@@ -1132,7 +1132,7 @@ void Test3D_decit(const Box<3,float> & domain, long int k)
 			// Grid sm
 			grid_sm<3,void> info(sz);
 
-			auto dom = g_dist.getSubDomainIterator({0,0,0},{sz[0]-2,sz[1]-2,sz[2]-2});
+			auto dom = g_dist.getSubDomainIterator({0,0,0},{(long int)sz[0]-2,(long int)sz[1]-2,(long int)sz[2]-2});
 
 			bool match = true;
